Add research scripts for debugging and verifying self-match and quantization processes
- Introduced 11_selfmatch_step_by_step_debug.py to analyze self-match discrepancies step-by-step.
- Created 12_manual_enroll_verify_ground_truth.py for manual reenactment of enroll and verify processes to check for consistency.
- Developed 13_real_handler_call_same_embedding.py to test handler.enroll() and handler.verify() directly on a known good embedding.
- Added 14_isolate_enroll_output_vs_verify_internals.py to isolate and compare outputs from the real verify function against manual calculations.
- Implemented 15_repeat_algebraic_identity_check.py to validate the algebraic identity of y_noisy_bits and codeword across multiple trials.
- Created 16_compare_real_enroll_vs_manual_same_seed.py to compare outputs of handler.enroll() with manual reenactment using fixed seeds.
- Added 17_check_confidence_scale.py to analyze the distribution of distance values and suggest appropriate scaling for SoftDistanceLLR.
- Developed 18_verify_quantizer_isolation.py to confirm that the new quantizer produces identical bits to the baseline quantizer across all embeddings.

+++ wifakey_module/wifakey_handler.py
@@ -252,7 +252,7 @@ class WiFaKeyHandler:
         b_selected = b_masked[:self.feature_length]
         y_noisy_bits = np.logical_xor(b_selected, helper_data)
 
-        y_llr = Modulation.BPSK(y_noisy_bits).reshape((1, self.N, self.Z))
+        y_llr = Modulation.BPSK(y_noisy_bits).astype(np.float32).reshape((1, self.N, self.Z)) # ép kiểu từ int64 sang float32
         y_pred_llr = self.sess.run(self.decoder_output, feed_dict={self.xa: y_llr})
 
         decoded_codeword = (y_pred_llr > 0).astype(int).flatten()
